Update dependencies and remove MonikerStore tests.
Update dependencies.  Tests now available in VS test runner.

Remove MonikerStore tests.

Making the MonikerStore tests work with XUnit would require refactoring the MonikerStore
class to accept the configuration dependency as an argument in the constructor.  This
would be A Good Thing, but it's not worth doing for a one-off project that will be
discarded after the new CMS in put in place.